Refactor: #185 폴더 구조 변경
