+++ PicCharge/PicCharge/Global/Helper/HapticManager.swift
