+++ PicCharge/PicCharge/Global/Utils/BatteryCalculator.swift
